CIV-0000 Enable R2 tests to run in AAT (#4743)
* enable R2 tests to run in AAT

* enable R2 tests to run in AAT

* update the tests

* update the tests

* update the tests

---------
