+++ src/test/functionalTests/tests/common/OCMC_tests.js
@@ -6,6 +6,7 @@ const CreateLipvLipClaimSteps  = require('../../citizenFeatures/createClaim/step
 // eslint-disable-next-line no-unused-vars
 let claimRef;
 
+// After CUI R2 release, OCMC tests are not needed to run in the pipeline. Hence added the ignore tag.
 Feature('Create Lip v Lip claim -  Full Admit and pay Immediately');
 
 Scenario('Create OCMC case and view it in preview @ocmc', async () => {
@@ -15,4 +16,4 @@ Scenario('Create OCMC case and view it in preview @ocmc', async () => {
   let claimRef = await CreateLipvLipClaimSteps.CreateClaimCreationOCMC(false);
   await LoginSteps.EnterCitizenCredentials(config.claimantCitizenUser.email, config.claimantCitizenUser.password);
   await CreateLipvLipClaimSteps.CheckOCMCcasePreview(claimRef);
-}).tag('@regression');
+}).tag('@ignore');
